refactor mobile banner
